move auto batch detection from gpt2 class to LM base class so we pad by longest document in task instead of maximum model CTX
